refactor: rename view class
