+++ app/src/main/res/layout/activity_list.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.list.ListActivity">
+    tools:context=".ui.questionlist.QuestionListActivity">
 
     <Button
         android:id="@+id/btnBack"
